+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## [0.1712.1](https://github.com/lightdash/lightdash/compare/0.1712.0...0.1712.1) (2025-06-12)
+
+
+### Bug Fixes
+
+* improve space movement with project scoping and cycle detection ([#15300](https://github.com/lightdash/lightdash/issues/15300)) ([5d658b4](https://github.com/lightdash/lightdash/commit/5d658b4ba5b44d55c389130187afbdd6a90be612))
+
+# [0.1712.0](https://github.com/lightdash/lightdash/compare/0.1711.0...0.1712.0) (2025-06-12)
+
+
+### Features
+
+* make ExportResults component wrappable in a modal ([#15303](https://github.com/lightdash/lightdash/issues/15303)) ([db2cdb3](https://github.com/lightdash/lightdash/commit/db2cdb3426c45c8f41471a396564fd759d97d7d0))
+* move create/edit agent to /ai-agents ([#15290](https://github.com/lightdash/lightdash/issues/15290)) ([61c90b3](https://github.com/lightdash/lightdash/commit/61c90b38936f80a3db5a0a73a76cc01f998a8015))
+
+# [0.1711.0](https://github.com/lightdash/lightdash/compare/0.1710.1...0.1711.0) (2025-06-12)
+
+
+### Features
+
+* Introduce Agent Switcher ([#15286](https://github.com/lightdash/lightdash/issues/15286)) ([6bdc31d](https://github.com/lightdash/lightdash/commit/6bdc31d376a6155f41f142b38a09009b96bffb2e)), closes [#15247](https://github.com/lightdash/lightdash/issues/15247)
+
 ## [0.1710.1](https://github.com/lightdash/lightdash/compare/0.1710.0...0.1710.1) (2025-06-12)